Al planificar la sincronización del cliente iOS me he encontrado con un detalle del modelo que iba a ser un problema a futuro. En el cliente local los puntos A y B del bucle se guardan como Double (los necesito así para poder marcar un loop con resolución por debajo del segundo, que es como trabajo), pero en el backend los tenía como int. Entonces al subir un bucle puesto en el segundo 12,4 el servidor lo iba a guardar como 12 y al volver a sincronizarlo al iPhone iba a perder esa precisión para siempre. Para arreglarlo he cambiado los campos puntoASegundos y puntoBSegundos a double tanto en la entidad JPA como en el DTO, y dejo que Hibernate aplique el ALTER TABLE solo gracias a ddl-auto=update. Aprovechando que es la primera vez que arranco el backend de verdad contra MySQL he arreglado dos cosas más que no estaban del primer día: añadir allowPublicKeyRetrieval=true a la URL JDBC y poner @Transactional en los métodos del CarpetaService que iteran sobre las canciones para calcular cantidadCanciones, porque sin transacción saltaba LazyInitializationException con open-in-view=false. He validado todo el round-trip con CURL, y el JSON de bucles llega y sale con decimales íntegros (12,345 → 12,345) y la lista de carpetas devuelve la cuenta correcta de canciones por carpeta.

+++ backend/src/main/java/com/loopmaster/backend/dto/BucleDTO.java
@@ -4,8 +4,8 @@ public class BucleDTO {
 
     private int id;
     private String nombre;
-    private int puntoASegundos;
-    private int puntoBSegundos;
+    private double puntoASegundos;
+    private double puntoBSegundos;
     private int velocidad;
     private int tonoSemitonos;
     private int cancionId;
@@ -13,7 +13,7 @@ public class BucleDTO {
     public BucleDTO() {
     }
 
-    public BucleDTO(int id, String nombre, int puntoASegundos, int puntoBSegundos, int velocidad, int tonoSemitonos, int cancionId) {
+    public BucleDTO(int id, String nombre, double puntoASegundos, double puntoBSegundos, int velocidad, int tonoSemitonos, int cancionId) {
         this.id = id;
         this.nombre = nombre;
         this.puntoASegundos = puntoASegundos;
@@ -39,19 +39,19 @@ public class BucleDTO {
         this.nombre = nombre;
     }
 
-    public int getPuntoASegundos() {
+    public double getPuntoASegundos() {
         return puntoASegundos;
     }
 
-    public void setPuntoASegundos(int puntoASegundos) {
+    public void setPuntoASegundos(double puntoASegundos) {
         this.puntoASegundos = puntoASegundos;
     }
 
-    public int getPuntoBSegundos() {
+    public double getPuntoBSegundos() {
         return puntoBSegundos;
     }
 
-    public void setPuntoBSegundos(int puntoBSegundos) {
+    public void setPuntoBSegundos(double puntoBSegundos) {
         this.puntoBSegundos = puntoBSegundos;
     }
